moved importFileModal to modals directory
