--- src/components/importFile/sections/descriptionArea/style.less
@@ -1,6 +1,0 @@
-@import '../../../../style/index';
-
-.title {
-    margin-bottom: 9px;
-}
-
