+++ src/components/modals/importFileModal/sections/descriptionArea/style.less
@@ -1,0 +1,6 @@
+@import '../../../../../style/index';
+
+.title {
+    margin-bottom: 9px;
+}
+
