+++ src/components/modals/importFileModal/sections/fileFinder/style.less
@@ -1,4 +1,4 @@
-@import '../../../../style/index';
+@import '../../../../../style/index';
 
 .container{
     height: 100%;
