+++ src/components/modals/importFileModal/style.less
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with Superblocks Lab.  If not, see <http://www.gnu.org/licenses/>.
 
-@import '../../style/index';
+@import '../../../style/index';
 
 .importModal {
     box-shadow: 0 2px 4px 0 rgba(0, 0, 0, 0.5);
